+++ lib/screens/my_page/make_profile_page.dart
@@ -8,7 +8,7 @@ import 'package:hexcolor/hexcolor.dart';
 import 'package:provider/provider.dart';
 
 import '../../providers/user_auth/authenticate.dart';
-import 'my_profile/my_profile_image.dart';
+import 'make_profile_image.dart';
 
 class MakeProfilePage extends StatefulWidget {
   final Stacks stacksProvider;
@@ -113,7 +113,7 @@ class _MakeProfilePageState extends State<MakeProfilePage> {
                   SizedBox(height: 20),
                   Container(color: Colors.grey),
                   Column(mainAxisAlignment: MainAxisAlignment.start, children: [
-                    MyProfileImage(size),
+                    MakeProfileImage(size),
                     _profileInfo(size, techStacks),
                     _authButton(size, setProfile)
                   ]),
